novdan_api/requirements.txt
asgiref==3.4.1
boto3==1.19.3
botocore==1.22.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.8
django-admin-rangefilter==0.8.1
django-cors-headers==3.10.0
django-oauth-toolkit==1.5.0
django-storages==1.12.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
jwcrypto==1.0
oauthlib==3.1.1
psycopg2==2.9.1
pycparser==2.20
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
wrapt==1.13.2
